Don't show live|recorded tabs for non-Replay recordings
